refactor: move app tls validation to services
