+++ api/types/app.go
@@ -17,9 +17,6 @@ limitations under the License.
 package types
 
 import (
-	"bytes"
-	"crypto/x509"
-	"encoding/pem"
 	"fmt"
 	"iter"
 	"net/url"
@@ -527,12 +524,6 @@ func (a *AppV3) CheckAndSetDefaults() error {
 		a.Metadata.Labels[AppSubKindLabel] = a.SubKind
 	}
 
-	if a.Spec.TLS != nil {
-		if err := a.checkTLS(); err != nil {
-			return trace.Wrap(err)
-		}
-	}
-
 	return nil
 }
 
@@ -635,11 +626,11 @@ func (a *AppV3) checkLLM() error {
 	// This also covers the requirement of supported format and provider values.
 	providers, ok := supportedFormatInferenceProviders[llm.Format]
 	if !ok {
-		return trace.BadParameter("Inference endpoint %q format %q doesn't have any valid 'provider'. Supported formats are: %s", a.GetName(), llm.Format, quoteAndJoin(SupportedLLMFormats))
+		return trace.BadParameter("Inference endpoint %q format %q doesn't have any valid 'provider'. Supported formats are: %s", a.GetName(), llm.Format, strings.Join(SupportedLLMFormats, ", "))
 	}
 
 	if !slices.Contains(providers, llm.Provider) {
-		return trace.BadParameter("Inference endpoint %q must set one of the providers supported by %q format: %s", a.GetName(), llm.Format, quoteAndJoin(providers))
+		return trace.BadParameter("Inference endpoint %q must set one of the providers supported by %q format: %s", a.GetName(), llm.Format, strings.Join(providers, ", "))
 	}
 
 	for _, model := range llm.Models {
@@ -658,119 +649,6 @@ func (a *AppV3) checkLLM() error {
 	return nil
 }
 
-func (a *AppV3) checkTLS() error {
-	if !AppSupportsTLSConfig(a.Spec.URI) {
-		return trace.BadParameter(
-			"App %q can only specify 'tls' settings for URI schemes that use upstream TLS. Supported schemes are: %s",
-			a.GetName(),
-			quoteAndJoin(appSchemasWithTLSSupport),
-		)
-	}
-
-	tls := a.Spec.TLS
-	var mode AppTLSMode
-	switch tls.Mode {
-	case AppTLSModeInsecure, AppTLSModeVerifyFull, AppTLSModeVerifyServerName, AppTLSModeVerifySpiffeID:
-		mode = tls.Mode
-	case "":
-		// When not specified, use the evaluated mode.
-		mode = a.GetTLSMode()
-	default:
-		return trace.BadParameter(
-			"App %q has invalid 'tls.mode' %q. Supported values are: %s",
-			a.GetName(),
-			tls.Mode,
-			quoteAndJoin([]string{AppTLSModeInsecure, AppTLSModeVerifyFull, AppTLSModeVerifyServerName, AppTLSModeVerifySpiffeID}),
-		)
-	}
-
-	if a.Spec.InsecureSkipVerify && mode != AppTLSModeInsecure {
-		return trace.BadParameter(
-			"App %q cannot specify 'insecure_skip_verify: true' (deprecated) and 'tls.mode: %q'. Drop 'insecure_skip_verify', and if you want the app to use insecure connections set 'tls.mode: %q'",
-			a.GetName(),
-			mode,
-			AppTLSModeInsecure,
-		)
-	}
-
-	switch tls.ClientCertMode {
-	case AppClientCertModeManaged:
-		if mode == AppTLSModeInsecure {
-			return trace.BadParameter("App %q can only enable 'tls.client_cert_mode' when 'tls.mode' is %q", a.GetName(), AppTLSModeVerifyFull)
-		}
-	case AppClientCertModeDisabled, "":
-	default:
-		return trace.BadParameter(
-			"App %q has invalid 'tls.client_cert_mode'. Supported values are: %s",
-			a.GetName(),
-			quoteAndJoin([]string{"", AppClientCertModeDisabled, AppClientCertModeManaged}),
-		)
-	}
-
-	switch mode {
-	case AppTLSModeVerifyFull:
-		// Note: tls.ServerName is optional and doesn't require any specific validation.
-		if !isValidSpiffeID(tls.ServerSpiffeId) {
-			return trace.BadParameter("App %q has invalid `tls.server_spiffe_id`. The SPIFFE ID must be complete (trust domain and path) and start with 'spiffe://'", a.GetName())
-		}
-	case AppTLSModeVerifyServerName:
-		// Note: tls.ServerName is optional and doesn't require any specific validation.
-		if tls.ServerSpiffeId != "" {
-			return trace.BadParameter("App %q 'tls.server_spiffe_id' is not used when mode is set to %q. To perform both, server name and SPIFFE ID verifications use %q mode", a.GetName(), mode, AppTLSModeVerifyFull)
-		}
-	case AppTLSModeVerifySpiffeID:
-		if !isValidSpiffeID(tls.ServerSpiffeId) {
-			return trace.BadParameter("App %q has invalid `tls.server_spiffe_id`. The SPIFFE ID must be complete (trust domain and path) and start with 'spiffe://'", a.GetName())
-		}
-		if tls.ServerName != "" {
-			return trace.BadParameter("App %q 'tls.server_name' is not used when mode is set to %q. To perform both, server name and SPIFFE ID verifications use %q mode", a.GetName(), mode, AppTLSModeVerifyFull)
-		}
-	case AppTLSModeInsecure:
-		if tls.ServerName != "" || tls.ServerSpiffeId != "" || len(tls.AllowedCas) > 0 {
-			return trace.BadParameter("App %q 'tls' are not in use since mode is set to %q", a.GetName(), mode)
-		}
-	}
-
-	for _, allowedCA := range tls.AllowedCas {
-		if slices.Contains(supportedAllowedInternalCAs, allowedCA) {
-			continue
-		}
-		if err := isValidSingleCertificatePEM(allowedCA); err != nil {
-			return trace.BadParameter(
-				"App %q 'tls.allowed_cas' values must each be a single PEM-encoded certificate or a Teleport CA alias (%s): %s",
-				a.GetName(),
-				quoteAndJoin(supportedAllowedInternalCAs),
-				err,
-			)
-		}
-	}
-
-	return nil
-}
-
-// isValidSingleCertificatePEM validates that s contains exactly one
-// PEM-encoded certificate.
-func isValidSingleCertificatePEM(s string) error {
-	block, rest := pem.Decode([]byte(s))
-	if block == nil {
-		return trace.BadParameter("expected a PEM-encoded certificate")
-	}
-	if _, err := x509.ParseCertificate(block.Bytes); err != nil {
-		return trace.BadParameter("invalid certificate: %s", err)
-	}
-	if len(bytes.TrimSpace(rest)) > 0 {
-		return trace.BadParameter("expected exactly one PEM-encoded certificate")
-	}
-	return nil
-}
-
-// isValidSpiffeID validates that s contains a valid SPIFFE ID.
-//
-// TODO(gabrielcorado): enhance the SPIFFE ID validation.
-func isValidSpiffeID(s string) bool {
-	return strings.HasPrefix(s, "spiffe://")
-}
-
 // AppSupportsTLSConfig returns if app supports TLS config based on its URI.
 func AppSupportsTLSConfig(uri string) bool {
 	parsed, err := url.Parse(uri)
@@ -778,20 +656,7 @@ func AppSupportsTLSConfig(uri string) bool {
 		return false
 	}
 
-	return slices.Contains(appSchemasWithTLSSupport, parsed.Scheme)
-}
-
-// quoteAndJoin takes a slice of strings and returns them quoted and
-// comma-separated.
-func quoteAndJoin(items []string) string {
-	if len(items) == 0 {
-		return ""
-	}
-	quotedItems := make([]string, len(items))
-	for i, item := range items {
-		quotedItems[i] = `"` + item + `"`
-	}
-	return strings.Join(quotedItems, ", ")
+	return slices.Contains(AppSchemesWithTLSSupport, parsed.Scheme)
 }
 
 // GetIdentityCenter returns the Identity Center information for the app, if any.
@@ -1013,8 +878,8 @@ var SupportedLLMProviders = []LLMProvider{
 	LLMProviderAWSBedrock,
 }
 
-// appSchemasWithTLSSupport list of app schemas that support TLS configurations.
-var appSchemasWithTLSSupport = []string{
+// AppSchemesWithTLSSupport list of app schemas that support TLS configurations.
+var AppSchemesWithTLSSupport = []string{
 	"https",
 	SchemeTLS,
 	SchemeMCPHTTPS,
@@ -1058,6 +923,6 @@ const (
 	AppTLSInternalCAWorkloadIdentity AppTLSInternalCA = "workload_identity"
 )
 
-// supportedAllowedInternalCAs is the list of internal CAs that can be used in
-// the app TLS configuration.
-var supportedAllowedInternalCAs = []string{AppTLSInternalCAWorkloadIdentity}
+// AppSupportedAllowedInternalCAs is the list of internal CAs that can be used
+// in the app TLS configuration.
+var AppSupportedAllowedInternalCAs = []string{AppTLSInternalCAWorkloadIdentity}
