+++ api/types/app_test.go
@@ -1046,197 +1046,6 @@ func TestLLMConfiguration(t *testing.T) {
 	})
 }
 
-func TestAppTLSConfiguration(t *testing.T) {
-	for name, tc := range map[string]struct {
-		uri          string
-		tls          *AppTLS
-		insecureSkip bool
-		expectErr    require.ErrorAssertionFunc
-	}{
-		"full valid configuration": {
-			tls: &AppTLS{
-				Mode:           AppTLSModeVerifyFull,
-				ServerName:     "example.com",
-				ServerSpiffeId: "spiffe://mycluster/svc/example",
-				AllowedCas:     []string{"workload_identity", appExternalAllowedCA},
-				ClientCertMode: AppClientCertModeManaged,
-			},
-			expectErr: require.NoError,
-		},
-		"mcp https valid configuration": {
-			uri: "mcp+https://localhost",
-			tls: &AppTLS{
-				Mode:           AppTLSModeVerifyFull,
-				ServerSpiffeId: "spiffe://mycluster/svc/mcp",
-				ClientCertMode: AppClientCertModeManaged,
-			},
-			expectErr: require.NoError,
-		},
-		"minimal": {
-			tls: &AppTLS{
-				Mode: AppTLSModeVerifyServerName,
-			},
-			expectErr: require.NoError,
-		},
-		"insecure with client cert mode": {
-			tls: &AppTLS{
-				Mode:           AppTLSModeInsecure,
-				ClientCertMode: AppClientCertModeManaged,
-			},
-			expectErr: require.Error,
-		},
-		"conflicting insecure skip verify and tls mode": {
-			tls: &AppTLS{
-				Mode: AppTLSModeVerifyServerName,
-			},
-			insecureSkip: true,
-			expectErr:    require.Error,
-		},
-		"incomplete server spiffe id": {
-			tls: &AppTLS{
-				Mode:           AppTLSModeVerifySpiffeID,
-				ServerSpiffeId: "/svc/example",
-			},
-			expectErr: require.Error,
-		},
-		"verify spiffe id mode and with server name": {
-			tls: &AppTLS{
-				Mode:           AppTLSModeVerifySpiffeID,
-				ServerName:     "example.com",
-				ServerSpiffeId: "spiffe://mycluster/svc/example",
-			},
-			expectErr: require.Error,
-		},
-		"invalid mode": {
-			tls: &AppTLS{
-				Mode: "invalid",
-			},
-			expectErr: require.Error,
-		},
-		"invalid allowed CA": {
-			tls: &AppTLS{
-				Mode:       AppTLSModeVerifyFull,
-				AllowedCas: []string{"workload_identity", "malformed cert"},
-			},
-			expectErr: require.Error,
-		},
-		"allowed CA with multiple certificates per entry": {
-			tls: &AppTLS{
-				Mode:       AppTLSModeVerifyFull,
-				AllowedCas: []string{appExternalAllowedCA + "\n" + appExternalAllowedCA},
-			},
-			expectErr: require.Error,
-		},
-		"server name with insecure mode": {
-			tls: &AppTLS{
-				Mode:       AppTLSModeInsecure,
-				ServerName: "example.com",
-			},
-			expectErr: require.Error,
-		},
-		"server spiffe id with insecure mode": {
-			tls: &AppTLS{
-				Mode:           AppTLSModeInsecure,
-				ServerSpiffeId: "spiffe://mycluster/svc/example",
-			},
-			expectErr: require.Error,
-		},
-		"invalid client cert mode value": {
-			tls: &AppTLS{
-				Mode:           AppTLSModeVerifyServerName,
-				ClientCertMode: "foo",
-			},
-			expectErr: require.Error,
-		},
-		"client cert mode with insecure": {
-			tls: &AppTLS{
-				Mode:           AppTLSModeInsecure,
-				ClientCertMode: AppClientCertModeManaged,
-			},
-			expectErr: require.Error,
-		},
-		"insecure skip verify compatible with mode insecure": {
-			tls:          &AppTLS{Mode: AppTLSModeInsecure},
-			insecureSkip: true,
-			expectErr:    require.NoError,
-		},
-		"insecure skip verify with empty mode defaults to insecure": {
-			tls:          &AppTLS{},
-			insecureSkip: true,
-			expectErr:    require.NoError,
-		},
-		"allowed cas with only internal alias": {
-			tls: &AppTLS{
-				Mode:       AppTLSModeVerifyServerName,
-				AllowedCas: []string{AppTLSInternalCAWorkloadIdentity},
-			},
-			expectErr: require.NoError,
-		},
-		"allowed cas with empty string entry": {
-			tls: &AppTLS{
-				Mode:       AppTLSModeVerifyServerName,
-				AllowedCas: []string{""},
-			},
-			expectErr: require.Error,
-		},
-		"allowed cas with non-certificate PEM block": {
-			tls: &AppTLS{
-				Mode:       AppTLSModeVerifyServerName,
-				AllowedCas: []string{appPrivateKeyPEM},
-			},
-			expectErr: require.Error,
-		},
-		"insecure mode with allowed cas": {
-			tls: &AppTLS{
-				Mode:       AppTLSModeInsecure,
-				AllowedCas: []string{appExternalAllowedCA},
-			},
-			expectErr: require.Error,
-		},
-		"cloud app with tls block": {
-			uri: "cloud://aws",
-			tls: &AppTLS{
-				Mode: AppTLSModeInsecure,
-			},
-			expectErr: require.Error,
-		},
-		"unsupported protocol with tls block": {
-			uri: "http://localhost",
-			tls: &AppTLS{
-				Mode: AppTLSModeInsecure,
-			},
-			expectErr: require.Error,
-		},
-		"unsupported protocol with empty tls config": {
-			uri:       "http://localhost",
-			tls:       &AppTLS{},
-			expectErr: require.Error,
-		},
-		"empty": {
-			tls:       &AppTLS{},
-			expectErr: require.NoError,
-		},
-		"nil": {
-			tls:       nil,
-			expectErr: require.NoError,
-		},
-	} {
-		t.Run(name, func(t *testing.T) {
-			uri := tc.uri
-			if uri == "" {
-				uri = "https://localhost"
-			}
-
-			_, err := NewAppV3(Metadata{Name: "myapp"}, AppSpecV3{
-				URI:                uri,
-				InsecureSkipVerify: tc.insecureSkip,
-				TLS:                tc.tls,
-			})
-			tc.expectErr(t, err)
-		})
-	}
-}
-
 func TestAppTLSMode(t *testing.T) {
 	for name, tc := range map[string]struct {
 		spec         AppSpecV3
@@ -1331,23 +1140,3 @@ func TestAppClientCertMode(t *testing.T) {
 		})
 	}
 }
-
-const (
-	appExternalAllowedCA = `-----BEGIN CERTIFICATE-----
-MIICFjCCAbugAwIBAgIRAIoq9H5/Uxz1oJS17hjFjKEwCgYIKoZIzj0EAwIwajEa
-MBgGA1UEChMRcm9vdC50ZWxlcG9ydC5kZXYxGjAYBgNVBAMTEXJvb3QudGVsZXBv
-cnQuZGV2MTAwLgYDVQQFEycxODM2NTY0OTg4MTY0NzM2OTQ4NzM2MzczNjQ4MTU2
-NjQzNTI0MTcwHhcNMjYwNDE2MjEwOTA4WhcNMzYwNDEzMjEwOTA4WjBqMRowGAYD
-VQQKExFyb290LnRlbGVwb3J0LmRldjEaMBgGA1UEAxMRcm9vdC50ZWxlcG9ydC5k
-ZXYxMDAuBgNVBAUTJzE4MzY1NjQ5ODgxNjQ3MzY5NDg3MzYzNzM2NDgxNTY2NDM1
-MjQxNzBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABDW7XPEplPkWWAZuQX6V/kqn
-T3iGcib+nYPFEOtWYZae3deF7wxcfyRTH4vsVOQhe9+TwJ92WwVnBCT/2U4QzgSj
-QjBAMA4GA1UdDwEB/wQEAwIBhjAPBgNVHRMBAf8EBTADAQH/MB0GA1UdDgQWBBSF
-51FfpZdVovLDB7CJg6BcU8MiFzAKBggqhkjOPQQDAgNJADBGAiEAh4qZv2xmwTM0
-TLezizSVzQRvrtY6t3lBqzSPUVx4JWcCIQDwBW5pscq6hxYBRpDRJxVghSobJiHk
-VdH2u/w+t7dV5g==
------END CERTIFICATE-----`
-	appPrivateKeyPEM = `-----BEGIN PRIVATE KEY-----
-MC4CAQAwBQYDK2VwBCIEIGp7zs9vCvGUPmVXFbgpppCQ5Tv7KFqGkV9vWzS1WInS
------END PRIVATE KEY-----`
-)
